Extract getParamsGroupByType
